fix(nextjs): add SVGR support for jest tests
