+++ docs/angular/api-next/builders/build.md
@@ -3,3 +3,35 @@
 Build a Next.js app
 
 Builder properties can be configured in angular.json when defining the builder, or when invoking it.
+
+## Properties
+
+### fileReplacements
+
+Type: `array` of `object`
+
+Replace files with other files in the build.
+
+#### replace
+
+Type: `string`
+
+undefined
+
+#### with
+
+Type: `string`
+
+undefined
+
+### outputPath
+
+Type: `string`
+
+The output path of the generated files.
+
+### root
+
+Type: `string`
+
+The source root
