+++ docs/react/api-next/builders/build.md
@@ -4,3 +4,35 @@ Build a Next.js app
 
 Builder properties can be configured in workspace.json when defining the builder, or when invoking it.
 Read more about how to use builders and the CLI here: https://nx.dev/react/guides/cli.
+
+## Properties
+
+### fileReplacements
+
+Type: `array` of `object`
+
+Replace files with other files in the build.
+
+#### replace
+
+Type: `string`
+
+undefined
+
+#### with
+
+Type: `string`
+
+undefined
+
+### outputPath
+
+Type: `string`
+
+The output path of the generated files.
+
+### root
+
+Type: `string`
+
+The source root
