+++ packages/next/src/schematics/application/lib/add-project.ts
@@ -1,0 +1,76 @@
+import { join, normalize } from '@angular-devkit/core';
+import { Rule } from '@angular-devkit/schematics';
+import { generateProjectLint } from '@nrwl/workspace';
+import { updateWorkspaceInTree } from '@nrwl/workspace/src/utils/ast-utils';
+import { NormalizedSchema } from './normalize-options';
+
+export function addProject(options: NormalizedSchema): Rule {
+  return updateWorkspaceInTree(json => {
+    const architect: { [key: string]: any } = {};
+    const { server } = options;
+
+    architect.build = {
+      builder: '@nrwl/next:build',
+      options: {
+        root: options.appProjectRoot,
+        outputPath: join(normalize('dist'), options.appProjectRoot)
+      },
+      configurations: {
+        production: {
+          fileReplacements: [
+            {
+              replace: `environments/environment.ts`,
+              with: `environments/environment.prod.ts`
+            }
+          ]
+        }
+      }
+    };
+
+    architect.serve = {
+      builder: '@nrwl/next:dev-server',
+      options: {
+        buildTarget: `${options.projectName}:build`,
+        dev: true
+      },
+      configurations: {
+        production: {
+          buildTarget: `${options.projectName}:build:production`,
+          dev: false
+        }
+      }
+    };
+
+    if (server) {
+      architect.serve.options = {
+        ...architect.serve.options,
+        customServerPath: options.server
+      };
+    }
+
+    architect.export = {
+      builder: '@nrwl/next:export',
+      options: {
+        buildTarget: `${options.projectName}:build:production`
+      }
+    };
+
+    architect.lint = generateProjectLint(
+      normalize(options.appProjectRoot),
+      join(normalize(options.appProjectRoot), 'tsconfig.json'),
+      options.linter
+    );
+
+    json.projects[options.projectName] = {
+      root: options.appProjectRoot,
+      sourceRoot: options.appProjectRoot,
+      projectType: 'application',
+      schematics: {},
+      architect
+    };
+
+    json.defaultProject = json.defaultProject || options.projectName;
+
+    return json;
+  });
+}
