+++ packages/next/src/schematics/application/lib/create-application-files.ts
@@ -1,0 +1,32 @@
+import {
+  apply,
+  filter,
+  mergeWith,
+  move,
+  noop,
+  Rule,
+  template,
+  url
+} from '@angular-devkit/schematics';
+import { names, offsetFromRoot } from '@nrwl/workspace';
+import { NormalizedSchema } from './normalize-options';
+
+export function createApplicationFiles(options: NormalizedSchema): Rule {
+  return mergeWith(
+    apply(url(`./files`), [
+      template({
+        ...names(options.name),
+        ...options,
+        tmpl: '',
+        offsetFromRoot: offsetFromRoot(options.appProjectRoot)
+      }),
+      options.styledModule
+        ? filter(file => !file.endsWith(`.${options.style}`))
+        : noop(),
+      options.unitTestRunner === 'none'
+        ? filter(file => file !== `/specs/index.spec.tsx`)
+        : noop(),
+      move(options.appProjectRoot)
+    ])
+  );
+}
